integer_ops: fix -Wformat warnings
The main sources of warnings were:

 * Printing of a `size_t` which requires the `%zu` specifier.

 * Printing of `cl_long`/`cl_ulong` which is now done using the
   `PRI*64` macros to ensure portability across 32 and 64-bit builds.

+++ test_conformance/integer_ops/test_popcount.cpp
@@ -20,6 +20,8 @@
 #include <sys/types.h>
 #include <sys/stat.h>
 
+#include <cinttypes>
+
 #include "procs.h"
 
 #define str(s) #s
@@ -51,7 +53,7 @@
             __popcnt(x, __T, _n, ref); \
             if(res != ref) \
             { \
-                log_info( "%ld) Failure for popcount( (%s%s) 0x%x ) = *%d vs %d\n", i, str(__T), sizeName, x, (int)ref, (int)res ); \
+                log_info( "%zu) Failure for popcount( (%s%s) 0x%x ) = *%d vs %d\n", i, str(__T), sizeName, (int)x, (int)ref, (int)res ); \
                 return -1; \
             }\
         } \
